debug command for dumping raw slab info
WARNING: This will hang the cache while it writes to disk!

Experimental debug feature. If it proves to be useful, I'll bake it behind a
configuration option and parameterize the filename, or have it write over TCP.

This *mostly* hangs the cache, in that things will shortly grind to a halt
but a few things in flight could modify data before being dumped. Does not
dump values.