test patch: malloc on set if client flags are "1"
this patch is for emulating a scenario where a multi-level cache exists
for persistent memory. The main slab memory should be on persistent
memory, and ranges of keys which are to be "cached" should be set with
a client flag of "1".

This has equivalent performance of using multiple slab allocators for
the purposes of benchmarking.