test patch: allow unix sockets and tcp connections
this is generally disallowed since users attempt to optimize pools by
changing the client lists on every machine, causing the caches to be
inconsistent. for this particular throughput test, I need to increase
the performance of a localhost benchmark running across a NUMA node. At
the same time TCP sockets are needed for more accurately measuring
relative latency overhead.

I should consider some way of force-enabling this in the future.