network: refactor binprot's bizarre rbuf usage
since keys were (maybe) expanding to 64k, binprot overloads the
read-into-key "conn_nread" state to read... a specific number of bytes
into the read buffer.

However keys were never expanded in size, so this ends up being a waste
of code and attaches binprot to management of the read buffer directly.

Instead, re-parse the binary header if we don't have enough bytes and
let try_read_network() handle it like it does for ascii.

As a side effect, this prevents multiple memmove's of potentially a
large amount of data for pipelined binprot commands, and under
NEED_ALIGN.