simplify meta command flag handling
flags are now space deliminated. tokens are gone. if a flag has an
argument, it comes attached to the flag, ie:

`mg foo q T30 c`

"quietly update the TTL to +30 and return the CAS value"

response codes are updated slightly. If a value is requested:
`VA [size] [flags]\r\nvalue\r\nEN\r\n`
if a value is not requested:
`HD [flags]\r\nEN\r\n`

code is shorter/simpler but it's a bigger change than I was hoping to
see at this point.