proxy: start of C side request object
lots of stubbed code, but this passes t/proxy.t now.
left a lot of WIP comments and some commented out code as well, please
excuse :)

have a good feeling for how much faster this would be vs keeping the
request object in lua. need to trace out the functions to check when
memory gets allocated, but that can wait.