proxy: BE error handling part one
not strictly tested yet, probably has bugs.

mostly plugging in the stubs. will now cancel the request stack on major
error and re-initiate connections. Part two will require separating the
connection logic to its own event; if able to connect and send a test
command, put backend into a normal state.

If the BE fails more than N times, mark it as bad and fast-fail requests
attempting to queue to it while a special timeout event wakes and
retries the above logic periodically.