proxy: untested jump hash code.
compiles and passes default tests, but haven't exercised the code yet.

to use jump hash the supplied server list must have a static order on
all participating servers. expansion and contraction of the pool must be
done by adding or removing servers to the end of the list. servers
within the list that die must be replaced in place.